(Possible Fix) Parade Hat Lex Aeterna Issue
Issue: Parade Hat is supposed to cast Lex Divina on the enemy target of
the Heal spell.  Instead, any time you case Heal, Lex Aeterna is cast on
the Priest himself.
Lex Divina is considered 18 (16 (A Buff) and 2 (Cannot be Saved)) when
it would make sense for it to be 34 (32: A Debuff, and 2 (Cannot be
saved).  This change my make the hat work as intended.