Switching to Renewal Weapon ASPD Penalty Values
Changing to these should improve ASPD values for all classes, although
has little affect on the required stats to reach the ASPD cap of 196 on
the server.  ASPD is heavily based on AGI and DEX/4.  Regardless of
weapon, you would need roughly the same stats to reach 196 aspd.  This
helps more in the earlier stages when AGI and DEX values are low but
less when they're higher.