Added Items related to Headgear NPC
Added numerous items that will be used by the headgear NPC for Middle
and Upper Headgears.  The NPC does not use these yet but will when the
NPC is coded further.
Added Item Group "Old_Golden_Box" for use with the "Box Stacker" Upper
Helmet Bonus.  Using one will currently only grant another "Old Golden
Box" until a list of items that can be thrown into this is made.

Note that generating any of these items in game should result in an
error as they would all need to be defined in the "iteminfo.lub" in the
system folder.