Headgear and Item Mall changes
Headgear Updates for Albums, Boxes, and Branches will now drop Album
Pouches, Box Pouches, and Branch Pouches.  These have a 1 in 50 chance
to drop a OCA/OPB/Bloody Branch and a 1 in 250 chance to drop an
Mystical Card Album/Old Gold Box/Nemesis Branch.  The remainder will
drop pieces which can be traded in when I make an NPC for it.

Removed items from the item mall that I had given to monsters previously
as drops.