Possible Fix for being unable to target homunculus with buffs
Removed the check in the conf regarding the homunculus being allowed or
not.  Also removed the "get master" check as well (wasn't sure if that
meant if the master were the one to grant the buff or not).

If i'm reading this right, it should be if the target is a Homunculus,
and if Skill_ID is not 0, and if the Skill ID is a support skill, then
return 1 and allow it to be cast.  As it was originally, provided that
line was false, it would've passed through and return 1 anyway, but
let's see what this does in any case.