Fixes to incorrect items effects, also Extended Job Max Level
Set Extended Job Levels maximum from 160 to 175.  Was set to 160 due to
the notion of "they are overpowered".  Needs testing before making that
suggestion and the end purpose is to balance all the classes, so can't
really test that while putting them at a 15 level handicap.

Fixed Freya Set's bonus to match its description (Says it should learn
Frost Nova and not Fire Pillar).

Fixed Freya Ring R to have its Heal and Teleport Level 1 as its
description states.

Fixed Wild Rider card to activate Level 1 Increase Agility when
attacking as its description says, not "when attacked" as its script was
before.

Wrote script for "CD in Mouth" as it is in its description.  Was not
programmed before.

Added "CD in Mouth" to Old Golden Boxes.